% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_add_deterministics_to_Z.R
\name{ci_add_deterministics_to_Z}
\alias{ci_add_deterministics_to_Z}
\title{Add deterministic terms to Z}
\usage{
ci_add_deterministics_to_Z(det_spec, q, Z)
}
\arguments{
\item{det_spec}{character, the deterministic specification}

\item{q}{numeric, the number of lags in the model}

\item{Z}{list, a list with Z0, Z1, Z2}
}
\value{
list with Z matrices updated with deterministics
}
\description{
Add deterministic terms to Z
}
