% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_add_exogenous_to_Z.R
\name{ci_add_exogenous_to_Z}
\alias{ci_add_exogenous_to_Z}
\title{Add exogenous variables to Z1 and Z2}
\usage{
ci_add_exogenous_to_Z(lags, Z, exogenous)
}
\arguments{
\item{lags, }{the number of lags in the model}

\item{Z}{list, the list with the Z matrices}

\item{exogenous}{xts object with the exogenous variables}
}
\value{
the list with Z0, Z1 and Z2
}
\description{
Add exogenous variables to Z1 and Z2
}
