% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_build_data_structures.R
\name{ci_build_data_structures}
\alias{ci_build_data_structures}
\title{Build basic data structures for estimation and testing.}
\usage{
ci_build_data_structures(data, spec = ci_spec(), make_R = TRUE,
  make_S = TRUE, exogenous = NULL)
}
\arguments{
\item{data}{an xts object, the series with the endogenous data.}

\item{spec}{a ci_spec object, the specification of the model.}

\item{make_R}{boolean, if TRUE; R matrices as discussed in Johansen(1996) 
are constructed.}

\item{make_S}{boolean, if TRUE; S matrices as discussed in Johansen(1996)
are constructed.}

\item{exogenous}{xts object, an time series with time index matching that
of data.}
}
\value{
a list with lists of Z, R and S matrices.
}
\description{
This function constructs the core data matrices used in
             in estimation and testing within the cointegrated vector
             error correction models. More specifically, the function
             builds the matrices Z0, Z1, Z2, R0, R1, S00, S10 and S11
             as these are described in 
             \insertCite{johansen_likelihood-based_1995}{civecm}.
}
\references{
{
  \insertRef{johansen_likelihood-based_1995}{civecm}
}
}
