% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calculate_omega.R
\name{ci_calculate_omega}
\alias{ci_calculate_omega}
\title{Calculate the error covariance matrix Omega}
\usage{
ci_calculate_omega(residuals)
}
\arguments{
\item{residuals, }{a T x p matrix of model residuals}
}
\value{
a matrix with the estimated error covariance
}
\description{
Calculate the error covariance matrix Omega
}
