% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calculate_residuals.R
\name{ci_calculate_residuals}
\alias{ci_calculate_residuals}
\title{Calculate the residuals in the model}
\usage{
ci_calculate_residuals(alpha, beta, R0, R1)
}
\arguments{
\item{alpha}{matrix, a p x r matrix with error correction paramters}

\item{beta}{matrix, a d x r matrix with long run relation, where d
is at least p and further depends on the deterministic 
specification and exogenous variables}

\item{R0}{matrix, a T x p matrix with concentrated residuals}

\item{R1}{matrix, a T x d matrix with concentrated residuals}
}
\value{
a T x p matrix with model residuals
}
\description{
Calculate the residuals in the model
}
