% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_estimate_unrestricted.R
\name{ci_estimate_unrestricted}
\alias{ci_estimate_unrestricted}
\title{Estimate the unrestricted model by Reduced Rank Regression}
\usage{
ci_estimate_unrestricted(R0, R1, r)
}
\arguments{
\item{R0}{matrix, as returned by the function ci_build_data_structures}

\item{R1}{matrix, as returned by the function ci_build_data_structures}

\item{r}{integer, the number of cointegration relations}
}
\value{
a list with an estimate for alpha and beta.
}
\description{
Estimate the unrestricted model by Reduced Rank Regression
}
