% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_fit.R
\name{ci_fit}
\alias{ci_fit}
\title{Fit a specified cointegration model to data using reduced rank regression}
\usage{
ci_fit(data, spec = ci_spec(), H_alpha = NULL, H_beta = NULL,
  h_beta = NULL, exogenous = NULL)
}
\arguments{
\item{data}{an xts object, time series data.}

\item{spec}{a ci_spec object, as returned by the function \code{ci_spec()} holding the
specification information on the model}

\item{H_alpha}{matrix, a matrix for imposing zero restrictions on alpha.}

\item{H_beta}{matrix, a design matrix with restrictions on the cointegration
relations.}

\item{h_beta}{vector, the design vector with values for the restricted values
in the cointegration relations}

\item{exogenous}{an xts object, time series data for included exogenous variables. These
could be more}
}
\value{
a ci_fit object
}
\description{
Fit a specified cointegration model to data using reduced rank regression
}
