% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_simulate.R
\name{ci_simulate}
\alias{ci_simulate}
\title{Simulate data based on a ci_vecm model}
\usage{
ci_simulate(model, nobs = 100, initial = NULL, shocks = NULL,
  exogenous = NULL)
}
\arguments{
\item{model}{a ci_vecm object, the model to simulate from.}

\item{nobs}{integer, the number of simulated observations. Ignored if
shocks or exogenous are provided.}

\item{initial}{xts object, a q x p xts object with initial values for the 
endogenous variables. If not provided (default is NULL), 
a matrix of zeroes q x p zeroes is used.}

\item{shocks}{xts object, nobs x p shocks to generate data.
Primarily usefull in bootstrapping. Default is NULL
in which case standard normal shocks are used together with 
Omega from the model object.}

\item{exogenous}{xts object, a (nobs + q) x pe matrix of exogenous variables 
that can be considered fixed. Here, q denotes the number of lags in 
the model. Mostly relevant for bootstrapping in a model with 
exogenous regressors. If shocks are provided, the number of rows 
in this matrix must match the number of rows in shocks plus 
the number of model lags.}
}
\value{
an xts object with with generated data
}
\description{
Simulate data based on a ci_vecm model
}
