% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_spec.R
\name{ci_spec}
\alias{ci_spec}
\title{Create a specification for a cointegrated vector autoregression}
\usage{
ci_spec(lags = 2, rank = 1, det_spec = "constant",
  seasonals = FALSE)
}
\arguments{
\item{lags}{numeric integer, number of lags in levels}

\item{rank}{numeric integer, the cointegration rank}

\item{det_spec}{character, the base deterministic specification. The ones
supported are the five given in Juselius(2006).
\itemize{
  \item none - no deterministics
  \item ci_constant - constant restricted to the cointegration relations
  \item constant - unrestricted constant
  \item ci_trend - unrestricted constant and trend restricted to the
        cointegration relations
  \item trend - unrestricted constant and trend
}}

\item{seasonals}{boolean, if TRUE seasonal dummies corresponding to the
data frequency will be included, monthly and quarterly are supported.}
}
\value{
creates an object of type ci_spec
}
\description{
Builds a specification for a standard cointegrated VAR model.
             It includes specifications of lag structure, cointegration rank
             and basic deterministic specifications.
}
