% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_switching_algorithm.R
\name{ci_switching_algorithm}
\alias{ci_switching_algorithm}
\title{The estimation switching algorithm to estimate the model under
       general linear restrictions}
\usage{
ci_switching_algorithm(unres_fit, S, H_alpha = NULL, H_beta = NULL,
  h_beta = NULL)
}
\arguments{
\item{unres_fit}{list, a list with unrestricted estimates for alpha and beta.}

\item{S}{list, a set of product moment matrices, S00, S01, S11 as 
described in e.g. Boswijk(1995) or Johansen(1995).}

\item{H_alpha}{matrix, a matrix for imposing zero restrictions on alpha.}

\item{H_beta}{matrix, a design matrix with restrictions on the cointegration
relations.}

\item{h_beta}{vector, the design vector with values for the restricted values
in the cointegration relations}
}
\value{
a list with the matrices alpha_hat, beta_hat and omega_hat
}
\description{
This function applies the method outlined in e.g. Boswijk(1995).
}
