% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_vecm.R
\name{ci_vecm}
\alias{ci_vecm}
\title{Create a cointegrated Vector Error Correction model object.}
\usage{
ci_vecm(spec, alpha, beta, Omega, Gamma = NULL, Phi = NULL)
}
\arguments{
\item{spec}{a ci_spec object, a ci_spec object defining the model specification.}

\item{alpha}{a matrix, the p x r matrix of error correction parameters.}

\item{beta}{a matrix, the matrix of long run relations which is at least p x r.}

\item{Omega}{a matrix, the positive semidefinite covariance matrix of the residuals.}

\item{Gamma}{a matrix, the short run parameters.}

\item{Phi}{a matrix, the coefficients for the exogenous variables outside the 
cointegration relations.}
}
\value{
ci_model object
}
\description{
Create a cointegrated Vector Error Correction model object.
}
