% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefmatrix.R
\name{coefmatrix}
\alias{coefmatrix}
\title{A function to print matrices of coefficient estimates with se og t.values}
\usage{
coefmatrix(coef, stat1, digits = 3, sig.dig = c(TRUE, FALSE))
}
\description{
A function to print matrices of coefficient estimates with se og t.values
}
