% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{eigen}
\alias{eigen}
\title{Calculates the eigenvalues of the companion matrix of the VAR.}
\usage{
eigen(object, ...)
}
\arguments{
\item{roots}{atomic vector or character, a vector of ranks or the default value 'max' 
which will give the eigenvalues for all ossible cointegration ranks.
It is possible parse the value 'specific' and the function will give the 
eigenvalues for the specific model with restrictions impoced.}
}
\value{
an object of class eigenCompanion
}
\description{
The values will be calculated without any restrictions on the parameters.
}
