% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null.R
\name{Null}
\alias{Null}
\title{Function borrowed from MASS but with here is the convention that the 
       Null of a zero column matrix is the identity}
\usage{
Null(M)
}
\description{
Function borrowed from MASS but with here is the convention that the 
       Null of a zero column matrix is the identity
}
